dbmate-tupac: a dbmate wrapper for better schemas
I have a problem with dbmate's schema handling. It will dump the entire
database "schema only" into a file, and expects that file to be useful
for reference. But much of the file can end up being changes due to
things outside the current project's concerns.

For example, enabling an extension that creates functions/tables in the
database will cause those to show up in the schema file. And things like
partitions that change over time: even if nothing changes in the
migrations, the dumped schema can show changes.

The "schema only" dump also excludes data that is effectively part of
the schema itself. For example, the `thread_state` table defines a set
of states that can be used for the `thread.status` field, and those rows
are a fixed set of values enforcing the set of possible `thread.status`
values. In effect, these "data" rows are as important as the info in a
"schema only" dump to ensuring the database has the information it needs
to work as expected.

As a result, it seems that the schema handling of dbmate does not give
us what we want, so I created this wrapper implementing a new `verify`
command that I think makes more sense. It does the following:

  - creates the database from the reference schema
  - dumps the database
  - drops the database
  - creates the database and applies all migrations
  - dumps the database
  - drops the database
  - diffs the dumps

If the diff shows no changes then the migrations and the reference
schema are aligned.

Functionally, this means that dbmate commands no longer dump the schema
into `schema.sql`, and we end up maintaining that file ourselves as the
true schema the migraitons must match.

This means that some changes will need to be made both in a migration
and in the reference schema. This is additional work. But it gives us a
way to actually enforce that the schema is an effective reference for
what we want to have in the database. In other words, this gives us a
reference schema that is not subject to the limitations of the "schema
only" dump supported by dbmate.