follow processes API spec for job results responses
If a job is not complete, we should return a 404 with a specific error
type indicating the reason the results are not found. We should also
return an appropriate error for result requests for finished
unsuccessful jobs; we're kinda doing that now, but it could use some
work.

This commit includes reworking the status code enums to centralize some
of the state sets, like terminal states, and the logic to determine if a
state instance is terminal.